+++ docs/utils/Id.mdx
@@ -1,5 +1,5 @@
 import { Meta } from '@storybook/addon-docs';
-import { useId } from '../../src/helpers/useId';
+import { useId } from '../../src/utils/useId';
 
 <Meta title="Docs/Utils/Id" />
