+++ src/form/Select/Select.tsx
@@ -13,7 +13,7 @@ import {
   ConnectedOverlayContentRef,
   Placement
 } from '../../utils';
-import { CloneElement, useId } from '../../helpers';
+import { CloneElement, useId } from '../../utils';
 import { SelectInput, SelectInputProps, SelectInputRef } from './SelectInput';
 import { SelectMenu, SelectMenuProps } from './SelectMenu';
 import { SelectOptionProps, SelectValue } from './SelectOption';
