+++ src/layers/Dialog/Dialog.tsx
@@ -1,7 +1,7 @@
 import React, { FC, ReactElement } from 'react';
 import classNames from 'classnames';
-import { GlobalOverlay, GlobalOverlayProps } from '../../utils';
-import { useId, CloneElement } from '../../helpers';
+import { GlobalOverlay, GlobalOverlayProps } from '../../utils/Overlay';
+import { useId, CloneElement } from '../../utils';
 import FocusTrap from 'focus-trap-react';
 import { motion } from 'framer-motion';
 import { twMerge } from 'tailwind-merge';
