+++ src/layers/Menu/Menu.tsx
@@ -2,7 +2,7 @@ import React, { FC, forwardRef, LegacyRef, useMemo } from 'react';
 import FocusTrap from 'focus-trap-react';
 import { ConnectedOverlay, OverlayEvent } from '../../utils/Overlay';
 import { Placement } from '../../utils/Position';
-import { useId } from '../../helpers';
+import { useId } from '../../utils';
 import { Modifiers } from 'popper.js';
 import { motion } from 'framer-motion';
 import { twMerge } from 'tailwind-merge';
